requirements.txt
flask==3.0.0
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.38